requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
Pygments==2.17.2
dnspython==2.4.2 